day 4 part 2

+++ day_four/day_four_test.go
@@ -20,7 +20,7 @@ Card 6: 31 18 13 56 72 | 74 77 10 23 35 67 36 11
 }
 
 func TestExampleTwo(t *testing.T) {
-	expectation := uint16(30)
+	expectation := uint32(30)
 	input := `
 Card 1: 41 48 83 86 17 | 83 86  6 31 17  9 48 53
 Card 2: 13 32 20 16 61 | 61 30 68 82 17 32 24 19
